:hammer: chore: minor fixes
